refactor: remove downloadRelayMessage function and related types, replacing it with getRelayMessages for improved message retrieval
