+++ core/mpc/relayMessage/RelayMessage.ts
@@ -1,0 +1,8 @@
+export type RelayMessage = {
+  session_id: string
+  from: string
+  to: string[]
+  body: string
+  hash: string
+  sequence_no: number
+}
