+++ core/mpc/schnorr/schnorrKeygen.ts
@@ -5,8 +5,6 @@ import {
   Keyshare,
   QcSession,
 } from '../../../lib/schnorr/vs_schnorr_wasm'
-import { deleteRelayMessage } from '../relayMessage/delete'
-import { downloadRelayMessage, RelayMessage } from '../downloadRelayMessage'
 import { waitForSetupMessage } from '../downloadSetupMessage'
 import {
   decodeDecryptMessage,
@@ -16,6 +14,8 @@ import { getKeygenThreshold } from '../getKeygenThreshold'
 import { getMessageHash } from '../getMessageHash'
 import { KeygenOperation } from '../keygen/KeygenOperation'
 import { initializeMpcLib } from '../lib/initialize'
+import { deleteRelayMessage } from '../relayMessage/delete'
+import { getRelayMessages } from '../relayMessage/get'
 import { combineReshareCommittee } from '../reshareCommittee'
 import { sendRelayMessage } from '../sendRelayMessage'
 import { sleep } from '../sleep'
@@ -107,12 +107,11 @@ export class Schnorr {
     const start = Date.now()
     while (true) {
       try {
-        const downloadMsg = await downloadRelayMessage({
+        const parsedMessages = await getRelayMessages({
           serverUrl: this.serverURL,
           localPartyId: this.localPartyId,
           sessionId: this.sessionId,
         })
-        const parsedMessages: RelayMessage[] = JSON.parse(downloadMsg)
         if (parsedMessages.length === 0) {
           // no message to download, backoff for 100ms
           await sleep(100)
