+++ lib/utils/query/queryUrl.ts
@@ -11,15 +11,21 @@ export function queryUrl(
   options: QueryUrlOptions & { responseType: 'none' }
 ): Promise<void>
 
-export function queryUrl<T>(
+export function queryUrl<T extends string = string>(
   url: string | URL,
-  options?: QueryUrlOptions & { responseType?: 'json' | 'text' }
+  options: QueryUrlOptions & { responseType: 'text' }
 ): Promise<T>
 
+export function queryUrl<T>(
+  url: string | URL,
+  options?: QueryUrlOptions & { responseType?: 'json' }
+): Promise<T>
+
+// Implementation
 export async function queryUrl<T>(
   url: string | URL,
   options: QueryUrlOptions = {}
-): Promise<T | void> {
+): Promise<T | string | void> {
   const { responseType = 'json', ...rest } = options
   const response = await fetch(url, {
     method: 'GET',
